refactor: genre 값 constant로 분리
