--- client/src/data/movie.js
@@ -1,124 +1,0 @@
-const MOVIE_DATA = [
-  {
-    id: 1,
-    title: 'Toy Story (1995)',
-    imdbid: '0114709',
-    tmdbid: '862',
-  },
-  {
-    id: 2,
-    title: 'Jumanji (1995)',
-    imdbid: '0113497',
-    tmdbid: '8844',
-  },
-  {
-    id: 3,
-    title: 'Grumpier Old Men (1995)',
-    imdbid: '0113228',
-    tmdbid: '15602',
-  },
-  {
-    id: 4,
-    title: 'Waiting to Exhale (1995)',
-    imdbid: '0114885',
-    tmdbid: '31357',
-  },
-  {
-    id: 5,
-    title: 'Father of the Bride Part II (1995)',
-    imdbid: '0113041',
-    tmdbid: '11862',
-  },
-  {
-    id: 6,
-    title: 'Heat (1995)',
-    imdbid: '0113277',
-    tmdbid: '949',
-  },
-  {
-    id: 7,
-    title: 'Sabrina (1995)',
-    imdbid: '0114319',
-    tmdbid: '11860',
-  },
-  {
-    id: 8,
-    title: 'Tom and Huck (1995)',
-    imdbid: '0112302',
-    tmdbid: '45325',
-  },
-  {
-    id: 9,
-    title: 'Sudden Death (1995)',
-    imdbid: '0114576',
-    tmdbid: '9091',
-  },
-  {
-    id: 10,
-    title: 'GoldenEye (1995)',
-    imdbid: '0113189',
-    tmdbid: '710',
-  },
-  {
-    id: 11,
-    title: 'American President, The (1995)',
-    imdbid: '0112346',
-    tmdbid: '9087',
-  },
-  {
-    id: 12,
-    title: 'Dracula: Dead and Loving It (1995)',
-    imdbid: '0112896',
-    tmdbid: '12110',
-  },
-  {
-    id: 13,
-    title: 'Balto (1995)',
-    imdbid: '0112453',
-    tmdbid: '21032',
-  },
-  {
-    id: 14,
-    title: 'Nixon (1995)',
-    imdbid: '0113987',
-    tmdbid: '10858',
-  },
-  {
-    id: 15,
-    title: 'Cutthroat Island (1995)',
-    imdbid: '0112760',
-    tmdbid: '1408',
-  },
-  {
-    id: 16,
-    title: 'Casino (1995)',
-    imdbid: '0112641',
-    tmdbid: '524',
-  },
-  {
-    id: 17,
-    title: 'Sense and Sensibility (1995)',
-    imdbid: '0114388',
-    tmdbid: '4584',
-  },
-  {
-    id: 18,
-    title: 'Four Rooms (1995)',
-    imdbid: '0113101',
-    tmdbid: '5',
-  },
-  {
-    id: 19,
-    title: 'Ace Ventura: When Nature Calls (1995)',
-    imdbid: '0112281',
-    tmdbid: '9273',
-  },
-  {
-    id: 20,
-    title: 'Money Train (1995)',
-    imdbid: '0113845',
-    tmdbid: '11517',
-  },
-]
-
-export { MOVIE_DATA }
