+++ client/src/routes/root.jsx
@@ -4,28 +4,7 @@ import { useNavigate } from 'react-router-dom'
 
 import BaseLayout from '../layouts/base-layout'
 import PopularCarousel from '../components/popular-carousel'
-
-const GENRES = [
-  { key: 'gen1', name: 'Adventure' },
-  { key: 'gen2', name: 'Animation' },
-  { key: 'gen3', name: 'Children' },
-  { key: 'gen4', name: 'Comedy' },
-  { key: 'gen5', name: 'Fantasy' },
-  { key: 'gen6', name: 'Romance' },
-  { key: 'gen7', name: 'Drama' },
-  { key: 'gen8', name: 'Action' },
-  { key: 'gen9', name: 'Crime' },
-  { key: 'gen10', name: 'Thriller' },
-  { key: 'gen11', name: 'Horror' },
-  { key: 'gen12', name: 'Mystery' },
-  { key: 'gen13', name: 'Sci-Fi' },
-  { key: 'gen14', name: 'War' },
-  { key: 'gen15', name: 'Musical' },
-  { key: 'gen16', name: 'Documentary' },
-  { key: 'gen17', name: 'IMAX' },
-  { key: 'gen18', name: 'Western' },
-  { key: 'gen19', name: 'Film-Noir' },
-]
+import { GENRES } from '../constants/genres'
 
 function Root() {
   const [title, setTitle] = useState('')
